moved singleton decorator to upper level to be used in other contexts
